Move hook #useIsTpmAvailable to a better place
