+++ web/src/hooks/model/storage/config-model.ts
@@ -23,8 +23,10 @@
 import { useCallback } from "react";
 import { useSuspenseQuery } from "@tanstack/react-query";
 import { useSystem } from "~/hooks/model/system/storage";
+import { useSystem as useBootloaderSystem } from "~/hooks/model/system/bootloader";
 import { solveStorageModel, getStorageModel, putStorageModel } from "~/api";
 import configModel from "~/model/storage/config-model";
+import bootloaderSystem from "~/model/system/bootloader";
 import { findDeviceByName } from "~/model/system/storage";
 import { isNullish } from "radashi";
 import type {
@@ -359,6 +361,16 @@ function useIsGrub2WithTpm(): boolean {
   return data;
 }
 
+function useIsTpmAvailable(): boolean {
+  const system = useBootloaderSystem();
+  const config = useConfigModel();
+  const bootloaderType = config ? configModel.getBootloader(config) : null;
+
+  if (isNullish(system) || isNullish(bootloaderType)) return false;
+
+  return bootloaderSystem.isTpmAvailable(system, bootloaderType);
+}
+
 export {
   STORAGE_MODEL_KEY,
   useConfigModel,
@@ -391,4 +403,5 @@ export {
   useSetSpacePolicy,
   useConvertDevice,
   useIsGrub2WithTpm,
+  useIsTpmAvailable,
 };
